fix(files_trashbin): Make files view registering an init script
