+++ apps/files_trashbin/lib/Listeners/LoadAdditionalScripts.php
@@ -37,6 +37,6 @@ class LoadAdditionalScripts implements IEventListener {
 			return;
 		}
 
-		Util::addScript(Application::APP_ID, 'main');
+		Util::addInitScript(Application::APP_ID, 'init');
 	}
 }
