+++ webpack.modules.js
@@ -72,7 +72,7 @@ module.exports = {
 		'personal-settings': path.join(__dirname, 'apps/files_sharing/src', 'personal-settings.js'),
 	},
 	files_trashbin: {
-		main: path.join(__dirname, 'apps/files_trashbin/src', 'main.ts'),
+		init: path.join(__dirname, 'apps/files_trashbin/src', 'files-init.ts'),
 	},
 	files_versions: {
 		files_versions: path.join(__dirname, 'apps/files_versions/src', 'files_versions_tab.js'),
@@ -81,7 +81,7 @@ module.exports = {
 		oauth2: path.join(__dirname, 'apps/oauth2/src', 'main.js'),
 	},
 	federatedfilesharing: {
-		'external': path.join(__dirname, 'apps/federatedfilesharing/src', 'external.js'),
+		external: path.join(__dirname, 'apps/federatedfilesharing/src', 'external.js'),
 		'vue-settings-admin': path.join(__dirname, 'apps/federatedfilesharing/src', 'main-admin.js'),
 		'vue-settings-personal': path.join(__dirname, 'apps/federatedfilesharing/src', 'main-personal.js'),
 	},
